+++ mybatis-spring-boot-starter/src/main/java/com/wkit/lost/mybatis/spring/boot/plugin/EnableInterceptors.java
@@ -1,6 +1,9 @@
 package com.wkit.lost.mybatis.spring.boot.plugin;
 
-import com.wkit.lost.mybatis.config.Plugin;
+import com.wkit.lost.mybatis.plugins.data.auditing.MetadataAuditingInterceptor;
+import com.wkit.lost.mybatis.plugins.locking.OptimisticLockingInterceptor;
+import com.wkit.lost.mybatis.plugins.paging.PageableInterceptor;
+import com.wkit.lost.mybatis.plugins.paging.RangePageableInterceptor;
 import org.springframework.context.annotation.Import;
 import org.springframework.core.annotation.AliasFor;
 
@@ -12,42 +15,47 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * 启用系统默认提供的插件
- * @see MyBatisInterceptorRegistrar
+ * 启用系统默认提供插件注解
+ * <p>
+ * <em>默认注册以下拦截器: </em>
+ *     <ul>
+ *         <li>元数据审计拦截器: {@link MetadataAuditingInterceptor}</li>
+ *         <li>乐观锁拦截器: {@link OptimisticLockingInterceptor}</li>
+ *         <li>常规分页拦截器: {@link PageableInterceptor}</li>
+ *         <li>范围分页拦截器: {@link RangePageableInterceptor}</li>
+ *     </ul>
+ * </p>
+ * @author wvkity
+ * @see MybatisInterceptorRegistrar
  */
-@Target( ElementType.TYPE )
-@Retention( RetentionPolicy.RUNTIME )
-@Import( { MyBatisInterceptorRegistrar.class } )
 @Documented
 @Inherited
+@Target( ElementType.TYPE )
+@Retention( RetentionPolicy.RUNTIME )
+@Import( { MybatisInterceptorRegistrar.class } )
 public @interface EnableInterceptors {
 
     /**
-     * 是否开启@Primary(默认为false)
-     * @return true: 是 false: 否
+     * 拦截器列表
+     * @return 拦截器列表
      */
-    @AliasFor( "primary" )
-    boolean value() default false;
+    @AliasFor( "interceptors" )
+    Plugin[] value() default {
+            @Plugin( value = MetadataAuditingInterceptor.class, order = 1 ),
+            @Plugin( value = OptimisticLockingInterceptor.class, order = 2 ),
+            @Plugin( value = PageableInterceptor.class, order = 3 ),
+            @Plugin( value = RangePageableInterceptor.class, order = 4 )
+    };
 
     /**
-     * 是否开启@Primary
-     * @return true: 是 false: 否
-     * @see org.springframework.context.annotation.Primary
+     * 拦截器列表
+     * @return 拦截器列表
      */
     @AliasFor( "value" )
-    boolean primary() default false;
-
-    /**
-     * 待注册插件列表
-     * @return 插件列表
-     */
-    Plugin[] include() default { Plugin.META_DATA_AUDIT, Plugin.OPTIMISTIC_LOCKING,
-            Plugin.PAGEABLE, Plugin.RANGE };
-
-    /**
-     * 过滤无需注册的插件
-     * @return 插件列表
-     */
-    Plugin[] exclude() default {};
-    
+    Plugin[] interceptors() default {
+            @Plugin( value = MetadataAuditingInterceptor.class, order = 1 ),
+            @Plugin( value = OptimisticLockingInterceptor.class, order = 2 ),
+            @Plugin( value = PageableInterceptor.class, order = 3 ),
+            @Plugin( value = RangePageableInterceptor.class, order = 4 )
+    };
 }